move common span exporting logic to @getstanza/sdk-base
