+++ packages/sdk-base/src/lib/init.test.ts
@@ -61,6 +61,8 @@ describe('Stanza init', function () {
 - apiKey: string (API key for a Hub instance)
 - serviceName: string (Name of the service)
 - serviceRelease: string (A version of the service)
+- sdkName: string (Name of the SDK)
+- sdkVersion: string (A version of the SDK)
 - environment: string (An environment to use)`);
     });
   });
