+++ packages/sdk-base/src/lib/open-telemetry/export/ManagedSpanProcessor.ts
